SD-2785-review comment changes
