+++ core/src/main/java/org/dcsa/conformance/core/party/ConformanceParty.java
@@ -2,6 +2,7 @@ package org.dcsa.conformance.core.party;
 
 import static org.dcsa.conformance.core.toolkit.JsonToolkit.OBJECT_MAPPER;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
@@ -398,5 +399,24 @@ public abstract class ConformanceParty implements StatefulEntity {
         statusCode,
         Map.of(API_VERSION, List.of(apiVersion)),
         new ConformanceMessageBody(OBJECT_MAPPER.createObjectNode().put("message", message)));
+  }
+
+  protected String getParamsForLogging(Map<String, Collection<String>> queryParams) {
+    Map<String, String> flattenedParams = flattenQueryParams(queryParams);
+    return toPrettyJson(flattenedParams);
+  }
+
+  private Map<String, String> flattenQueryParams(Map<String, Collection<String>> queryParams) {
+
+    return queryParams.entrySet().stream()
+        .collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().iterator().next()));
+  }
+
+  private String toPrettyJson(Map<String, String> map) {
+    try {
+      return OBJECT_MAPPER.writeValueAsString(map);
+    } catch (JsonProcessingException e) {
+      throw new RuntimeException("Failed to serialize query params", e);
+    }
   }
 }
